fix weighted model retrieving
Yes, that `.toInt()` was missing. It literally took me over one hour to debug. I attached a debugger to minecraft and it was a total mess.

+++ src/main/java/de/bixilon/minosoft/gui/rendering/models/block/state/render/WeightedBlockRender.kt
@@ -40,7 +40,7 @@ class WeightedBlockRender(
         if (random == null) return models.first().model
         random.setSeed(position.positionHash)
 
-        var weightLeft = abs(random.nextLong() % totalWeight)
+        var weightLeft = abs(random.nextLong().toInt() % totalWeight)
 
         for ((weight, model) in models) {
             weightLeft -= weight
